+++ changelog.md
@@ -1,4 +1,9 @@
 # Change log
+### [1.2.3] - 2020-02-14
+#### Changes
+- Added ability to remove page-wide proxy
+- Changed static classes to object literals for compability with Node.js **10.16.x** ([#6](https://github.com/Cuadrix/puppeteer-page-proxy/issues/6))
+- Removed `src\util\` folder along with proxy-validator
 ### [1.2.2] - 2020-02-09
 #### Patches
 - Fixed code indentation.